Cast estimated_num_rows to uint64_t.

+++ src/include/tscout/marker.h
@@ -26,12 +26,12 @@
 #define TS_MARKER_SETUP()               \
   /* Features. */                       \
   uint64_t query_id;                    \
-  double estimated_num_rows;            \
+  uint64_t estimated_num_rows;          \
   int estimated_row_width_bytes;
 
 // Define common features.
-#define TS_FEATURES_MARKER(name, plan_state_ptr, ...) \
-  query_id = plan_state_ptr->state->es_plannedstmt->queryId; \
-  estimated_num_rows = plan_state_ptr->plan->plan_rows; \
-  estimated_row_width_bytes = plan_state_ptr->plan->plan_width; \
+#define TS_FEATURES_MARKER(name, plan_state_ptr, ...)                                           \
+  query_id = plan_state_ptr->state->es_plannedstmt->queryId;                                    \
+  estimated_num_rows = (double) plan_state_ptr->plan->plan_rows;                                \
+  estimated_row_width_bytes = plan_state_ptr->plan->plan_width;                                 \
   TS_MARKER(name, query_id, estimated_num_rows, estimated_row_width_bytes, ##__VA_ARGS__);
